Continued Work on CustomizationCheck -A blank page appears that should, but does not, link to the customization page.
