--- app/src/main/java/com/example/testing/ui/customizationTab/CustomizationCheckFragment.java
@@ -1,29 +1,0 @@
-package com.example.testing.ui.customizationTab;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-
-import com.example.testing.databinding.FragmentNothomeBinding;
-
-public class CustomizationCheckFragment extends Fragment{
-
-    private FragmentNothomeBinding binding;
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-
-        binding = FragmentNothomeBinding.inflate(inflater, container, false);
-
-        return binding.getRoot();
-
-    }
-    @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        binding = null;
-    }
-}
