+++ app/src/main/res/layout/fragment_customizationcheck.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notHome.NotHomeFragment">
+    tools:context=".ui.customizationTab.CustomizationFragment">
 
     <TextView
-        android:id="@+id/text_notHome"
+        android:id="@+id/text_CustomizationCheck"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
@@ -21,7 +21,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <Button
-        android:id="@+id/button"
+        android:id="@+id/customizationButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="158dp"
